packages/backend/pyproject.toml
[tool.poetry]
name = "nova-backend"
version = "0.1.0"
description = "The Universal Trust Layer for AI - Backend Service"
authors = ["Sanjana <ranasanjana60@gmail.com>", "Rama Pranav <gantiramapranav@gmail.com>"] 
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10, <4.0"
openai = ">=1.97.1,<2.0.0"
pydantic = ">=2.11.7,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"

fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlalchemy = "^2.0.31"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.4"
qdrant-client = "^1.9.0"
langchain = "^0.2.11"
langchain-openai = "^0.1.15"
langchain-community = "^0.2.11"
langchain-core = "^0.2.14"
tiktoken = "^0.7.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
